iter1-auth: Refactoring app (construct)
